Add files via upload
These are more of the labs from CSA. I uploaded these in no specific order, and I didn't upload all of the labs (especially the basic ones or the ones that require obscure packages like gpdraw to run). They should give a good idea of what was done in the course. I recieved A's on all of them (and in the class).